removes any other query parameters than essential ?v=
 The youtube video that is revisited may has additional query
parameter to resume the video from location where the user has
watched up to previsouly. This parameter will however make
youtube-dl to fail to retrieve information somehow. So in that
case, The video url parameter has to be edited to remove the
parameter manually which is not a hard task but it would be ideal
and convenient for this tool to do it by itself automatically.